% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{bed2GRanges}
\alias{bed2GRanges}
\title{bed2GRanges : read bed to GRanges}
\usage{
bed2GRanges(BEDFile, header = FALSE)
}
\arguments{
\item{BEDFile}{bed file}
}
\value{
bed information, in GRanges object
}
\description{
bed2GRanges : read bed to GRanges
}
\examples{
BEDFile="./test_data/E094_DNase.all.peaks.v2.bed"
b2G=bed2GRanges(BEDFile,header = FALSE)
head(b2G)
}
