% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fimo_to_GRanges.R
\name{filter_fimo}
\alias{filter_fimo}
\title{filter fimo_CTCF.}
\usage{
filter_fimo(dnase, fimo_GRanges)
}
\arguments{
\item{dnase}{granges objects}

\item{fimo_GRanges}{ctcf_motif granges objects}
}
\value{
filtered fimo GRanges objects
}
\description{
filter fimo_CTCF.
}
\examples{

}
