% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fimo_to_GRanges.R
\name{filter_fimo_to_1000bp_gr}
\alias{filter_fimo_to_1000bp_gr}
\title{extension 1000bp about center motif.}
\usage{
filter_fimo_to_1000bp_gr(motif.ovl)
}
\arguments{
\item{motif.ovl}{}
}
\value{
site 1000 bp region
}
\description{
extension 1000bp about center motif.
}
\examples{

}
