% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Fimo2GRanges}
\alias{Fimo2GRanges}
\title{Title}
\usage{
Fimo2GRanges(fimo_file, as = "GRanges", ...)
}
\arguments{
\item{fimo_file}{FIMO_TSV.file}

\item{as}{output type}

\item{...}{other}
}
\value{
data_strucure such Grange or bed
}
\description{
Title
}
