% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fimo_to_GRanges.R
\name{mut_pos_tb}
\alias{mut_pos_tb}
\title{output position in 1000bp flank region.}
\usage{
mut_pos_tb(site, somatic_gastric)
}
\arguments{
\item{site}{flank-region granges objects}

\item{somatic_gastric}{granges objects}
}
\value{
mutation_pos_table
}
\description{
output position in 1000bp flank region.
}
\examples{

}
