% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flank_5bp_loci.R
\name{plot_flank_5bp}
\alias{plot_flank_5bp}
\title{plot mutation direction in the region 0f 5bp}
\usage{
plot_flank_5bp(motif.ovl, somatic_COLO829)
}
\arguments{
\item{motif.ovl}{oberlaped motif granges}

\item{somatic_COLO829}{snp Granges}
}
\value{
graph
}
\description{
plot mutation direction in the region 0f 5bp
}
\examples{

}
