% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutation_1k_loci.R
\name{plot_mutation_loci}
\alias{plot_mutation_loci}
\title{1000bp mutation loci}
\usage{
plot_mutation_loci(
  mut.table,
  center_expend_each = 5,
  tf_name = "NA",
  mutationScore_outdir = "./"
)
}
\arguments{
\item{mut.table}{mutation position}

\item{center_expend_each}{expand center region to statictic enrichment}

\item{tf_name}{Transcription factor name}

\item{mutationScore_outdir}{enrichment score output directory}
}
\value{

}
\description{
1000bp mutation loci
}
\examples{

}
